requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.2
Flask_Cors==4.0.0
joblib==1.2.0
pandas==2.0.2
python_whois==0.8.0
Requests==2.31.0
scikit_learn==1.2.2
tldextract==3.4.4
